requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]>=4.7,<5
scikit-learn
dill
pyaml
dagshub
mlflow
##-e .